Make MP games between different versions (e.g. Win 64 and 32 bit) compatible again

+++ ctp2_code/gs/slic/slicif.h
@@ -202,7 +202,7 @@ extern int g_slicWarnUndeclared;
 extern FILE *debuglog;
 
 unsigned char* slicif_get_code();
-int slicif_get_code_size();
+size_t slicif_get_code_size();
 SLIC_ERROR slicif_run_parser(char *filename, int symStart);
 void slicif_add_object(struct PSlicObject* obj);
 void slicif_add_op(SOP op, ...);
